requirements.txt
flask==3.0.0
PyMuPDF==1.24.5
pytesseract==0.3.10
Pillow==10.3.0
gunicorn==21.2.0
